requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
apimatic-core==0.2.19
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
asgiref==3.8.1
attrs==25.1.0
CacheControl==0.12.14
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
dateparser==1.2.1
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-picklefield==3.2
django-q2==1.7.6
django-sslserver==0.22
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.8.2
jsonpickle==3.3.0
jsonpointer==2.4
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
openai==1.65.5
packaging==24.2
pillow==11.1.0
postgres==4.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
retell-sdk==4.19.0
rsa==4.9
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
square==0.0.3
squareup==40.1.0.220250220
tqdm==4.67.1
twilio==9.4.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.18.3
